@@ -11,7 +11,7 @@ keywords = ["gdpr", "pseudonymization", "nlp", "privacy", "data-protection"]
 packages = [{include = "gdpr_pseudonymizer"}]
 
 [tool.poetry.dependencies]
-python = ">=3.9.2,<3.14"
+python = ">=3.10,<3.14"
 spacy = "^3.7.0"
 typer = "^0.9.0"
 rich = "^13.7.0"
@@ -27,7 +27,7 @@ pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
 pytest-mock = "^3.12.0"
 pytest-benchmark = "^4.0.0"
-black = "^23.12.0"
+black = "^24.3.0"
 ruff = "^0.1.0"
 mypy = "^1.7.0"
 stanza = "^1.7.0"
@@ -46,7 +46,7 @@ build-backend = "poetry.core.masonry.api"
 # Black configuration
 [tool.black]
 line-length = 88
-target-version = ['py39']
+target-version = ['py310']
 include = '\.pyi?$'
 extend-exclude = '''
 /(
@@ -65,7 +65,7 @@ extend-exclude = '''
 # Ruff configuration
 [tool.ruff]
 line-length = 88
-target-version = "py39"
+target-version = "py310"
 exclude = [
     ".bzr",
     ".direnv",
@@ -99,15 +99,13 @@ select = [
 ]
 ignore = [
     "E501",  # Line too long (handled by Black)
-    "UP007",  # Use X | Y for type annotations - SQLAlchemy Mapped[] requires Optional[] for Python 3.9
-    "UP037",  # Remove quotes from type annotation (needed for TYPE_CHECKING imports)
 ]
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]  # Allow unused imports in __init__.py
 "scripts/*.py" = ["E402"]  # Allow imports after sys.path modification
 "tests/unit/test_*.py" = ["E402"]  # Allow imports after sys.path modification
-# Note: models.py uses Optional[] in Mapped[] for SQLAlchemy Python 3.9 compatibility
+"gdpr_pseudonymizer/cli/**/*.py" = ["UP007"]  # Typer 0.9.x requires Optional[] syntax
 
 # pytest configuration
 [tool.pytest.ini_options]
@@ -157,7 +155,7 @@ exclude_lines = [
 
 # mypy configuration
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
